refactor: 클라이언트 소켓 읽기, 쓰기 클래스 정의
- 클라이언트 소켓 대신 클라이언트 소켓의 읽기, 쓰기 기능에만 접근 가능하도록 클라이언트 소켓의 읽기, 쓰기 클래스 정의
